# docs/api-contract.yaml
openapi: 3.1.0
info:
  title: Ride-hailing API (Baghdad v1)
  version: 1.0.0
  description: |
    Binding API contract. CLAUDE.md §12.1: no endpoint may exist in the code that
    is not in this file. Adding an endpoint means changing this file first, in
    its own commit, with a reason.

    Conventions:
      * All money fields are named `*_iqd`, are integers, and are whole Iraqi
        Dinars (CLAUDE.md §6.1). A fractional value in any `*_iqd` field is a
        contract violation, not a rounding detail.
      * All timestamps are RFC 3339 UTC. Clients render in Asia/Baghdad.
      * All errors are RFC 9457 `application/problem+json` (CLAUDE.md §9).
      * Every request carries `X-Request-Id`; the server generates one if absent
        and always echoes it.
      * Bearer JWT auth unless the operation is explicitly marked `security: []`.
servers:
  - url: /v1
...
security:
  - bearerAuth: []
paths:
...
  /fare/estimate:
    post:
      tags: [rider]
      operationId: estimateFare
      summary: |
        Quote a fare. Pure arithmetic over the configured tariff plus a routing
        distance. The routing call is NOT made synchronously here (CLAUDE.md
        §3.2) — the server uses a haversine estimate with a configured road
        factor. Precise routing is a queued job used only for post-trip
        reconciliation.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [pickup, dropoff]
              properties:
                pickup:   { $ref: '#/components/schemas/LatLng' }
                dropoff:  { $ref: '#/components/schemas/LatLng' }
      responses:
        '200':
          description: Quote.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FareEstimate' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
  /rides:
    post:
      tags: [rider]
      operationId: createRide
      summary: |
        Request a ride. CLAUDE.md §5.2: `Idempotency-Key` is REQUIRED. Replaying
        the same key with the same body within 24h returns the ORIGINAL ride
        with HTTP 200 and never creates a second ride.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [pickup, dropoff]
              properties:
                pickup:          { $ref: '#/components/schemas/LatLng' }
                pickupAddress:   { type: string, maxLength: 400 }
                dropoff:         { $ref: '#/components/schemas/LatLng' }
                dropoffAddress:  { type: string, maxLength: 400 }
                paymentMethod:
                  type: string
                  enum: [CASH]
                  default: CASH
                  description: GATEWAY is defined in the schema but rejected in v1 (CLAUDE.md §7).
                proposedFareIqd:
                  type: integer
                  minimum: 1
                  maximum: 100000000
                  description: |
                    What the rider offers to pay, in whole IQD. Optional, and
                    ignored entirely unless `negotiation_enabled` is on.

                    Supplying it is what opens the ride to driver bids — without
                    it `/rides/{rideId}/bids` has nothing to bid against, which
                    is why bidding was unreachable through the API before this
                    field existed.

                    Bounded server-side to `negotiation_band_bps` around the
                    metered estimate. A rider must not be able to anchor at 500
                    IQD, and a driver must not be able to hold out for ten times
                    the meter.
      responses:
        '201':
          description: New ride created and queued for matching.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Ride' }
        '200':
          description: |
            Idempotency-Key replay. This is the ORIGINAL ride, unchanged.
            Distinguishable from 201 by status code alone.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Ride' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: |
            Either the rider already has a live ride, or the same
            Idempotency-Key was reused with a DIFFERENT body.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '422':
          $ref: '#/components/responses/ValidationError'
  /rides/me:
    get:
      tags: [shared]
      operationId: listMyRides
      summary: |
        The caller's own rides. A rider sees rides where they are the rider; a
        driver sees rides where they are the assigned driver. There is no
        parameter that widens this — the filter is derived from the token, not
        from input (ACCEPTANCE_CHECKLIST.md check 5).
      parameters:
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/RideStatus' }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 50, default: 20 }
        - name: cursor
          in: query
          schema: { type: string, maxLength: 200 }
      responses:
        '200':
          description: Page of rides, newest first.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RidePage' }
        '401':
          $ref: '#/components/responses/Unauthorized'
  /rides/{rideId}:
    parameters:
      - $ref: '#/components/parameters/RideId'
    get:
      tags: [shared]
      operationId: getRide
      summary: |
        A single ride. Readable only by its rider, its assigned driver, a driver
        holding a live offer on it, or an admin. Everyone else gets 404 — not
        403 — so the endpoint does not confirm that a ride id exists.
      responses:
        '200':
          description: The ride.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Ride' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /rides/{rideId}/cancel:
    parameters:
      - $ref: '#/components/parameters/RideId'
    post:
      tags: [shared]
      operationId: cancelRide
      summary: |
        Cancel. The resulting state depends on the caller's role:
        rider → CANCELLED_BY_RIDER, driver → CANCELLED_BY_DRIVER,
        admin on an IN_PROGRESS ride → CANCELLED_IN_TRIP.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: { type: string, maxLength: 400 }
      responses:
        '200':
          description: Cancelled.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Ride' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/InvalidTransition'
  /rides/{rideId}/rate:
    parameters:
      - $ref: '#/components/parameters/RideId'
    post:
      tags: [shared]
      operationId: rateRide
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [score]
              properties:
                score:   { type: integer, minimum: 1, maximum: 5 }
                comment: { type: string, maxLength: 1000 }
      responses:
        '201':
          description: Rating stored.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Rating' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Ride is not COMPLETED, or the caller already rated it.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '422':
          $ref: '#/components/responses/ValidationError'

  # ===========================================================================
  # DRIVER
  # ===========================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    RideId:
      name: rideId
      in: path
      required: true
      schema: { type: string, format: uuid }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Client-generated UUID v4. CLAUDE.md §5.2.
      schema: { type: string, format: uuid }
  responses:
    Unauthorized:
      description: Missing or invalid token.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Forbidden:
      description: Authenticated but not permitted.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: Not found, or not visible to this caller.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    InvalidTransition:
      description: The requested ride state transition is not permitted from the current state.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: https://api.rideapp.iq/problems/invalid-ride-transition
            title: Invalid ride transition
            status: 409
            detail: Cannot move ride from COMPLETED to IN_PROGRESS.
            fromState: COMPLETED
            toState: IN_PROGRESS
    ValidationError:
      description: Request failed schema validation.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
...
    Problem:
      type: object
      description: RFC 9457 problem details.
      required: [type, title, status]
      properties:
        type:     { type: string, format: uri }
        title:    { type: string }
        status:   { type: integer }
        detail:   { type: string }
        instance: { type: string }
        requestId: { type: string }
        errors:
          type: array
          description: Field-level validation failures.
          items:
            type: object
            required: [path, message]
            properties:
              path:    { type: string }
              message: { type: string }
      additionalProperties: true
    LatLng:
      type: object
      required: [lat, lng]
      properties:
        lat: { type: number, minimum: -90,  maximum: 90 }
        lng: { type: number, minimum: -180, maximum: 180 }
...
    RideStatus:
      type: string
      enum:
        - REQUESTED
        - OFFERED
        - ACCEPTED
        - DRIVER_ARRIVED
        - IN_PROGRESS
        - COMPLETED
        - CANCELLED_BY_RIDER
        - CANCELLED_BY_DRIVER
        - CANCELLED_IN_TRIP
        - EXPIRED
        - NO_DRIVERS_FOUND
...
    PublicUser:
      type: object
      description: |
        How a counterparty is shown to the other side of a ride. Deliberately
        has NO phone field. ACCEPTANCE_CHECKLIST.md check 5 requires that a
        driver cannot see the rider's number after the trip; the way this is
        guaranteed is that the field does not exist in the response shape at all.
      required: [id, displayName]
      properties:
        id:          { type: string, format: uuid }
        displayName: { type: string }
        rating:      { type: number, nullable: true, minimum: 1, maximum: 5 }
        vehicle:
          type: object
          nullable: true
          properties:
            plate: { type: string }
            model: { type: string }
            color: { type: string }
...
    FareEstimate:
      type: object
      required: [estimatedFareIqd, distanceM, durationS, breakdown]
      properties:
        estimatedFareIqd: { type: integer, minimum: 0, description: Whole IQD. }
        distanceM:        { type: integer, minimum: 0 }
        durationS:        { type: integer, minimum: 0 }
        breakdown:
          type: object
          required: [baseIqd, distanceIqd, timeIqd, minimumAppliedIqd, roundingIqd]
          properties:
            baseIqd:           { type: integer, minimum: 0 }
            distanceIqd:       { type: integer, minimum: 0 }
            timeIqd:           { type: integer, minimum: 0 }
            minimumAppliedIqd: { type: integer, minimum: 0 }
            roundingIqd:       { type: integer, minimum: 0 }
    Ride:
      type: object
      required: [id, status, pickup, dropoff, estimatedFareIqd, paymentMethod, requestedAt]
      properties:
        id:     { type: string, format: uuid }
        status: { $ref: '#/components/schemas/RideStatus' }
        rider:  { $ref: '#/components/schemas/PublicUser' }
        driver:
          oneOf:
            - $ref: '#/components/schemas/PublicUser'
            - type: 'null'
        pickup:          { $ref: '#/components/schemas/LatLng' }
        pickupAddress:   { type: string, nullable: true }
        dropoff:         { $ref: '#/components/schemas/LatLng' }
        dropoffAddress:  { type: string, nullable: true }
        estimatedFareIqd:   { type: integer, minimum: 0 }
        finalFareIqd:       { type: integer, nullable: true, minimum: 0 }
        commissionIqd:      { type: integer, nullable: true, minimum: 0 }
        estimatedDistanceM: { type: integer, minimum: 0 }
        estimatedDurationS: { type: integer, minimum: 0 }
        actualDistanceM:    { type: integer, nullable: true, minimum: 0 }
        paymentMethod:      { type: string, enum: [CASH, GATEWAY] }
        requestedAt:        { type: string, format: date-time }
        acceptedAt:         { type: string, format: date-time, nullable: true }
        driverArrivedAt:    { type: string, format: date-time, nullable: true }
        startedAt:          { type: string, format: date-time, nullable: true }
        completedAt:        { type: string, format: date-time, nullable: true }
        cancelledAt:        { type: string, format: date-time, nullable: true }
        cancellationReason: { type: string, nullable: true }
    RidePage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Ride' }
        nextCursor: { type: string, nullable: true }
...
    Rating:
      type: object
      required: [id, rideId, score, createdAt]
      properties:
        id:        { type: string, format: uuid }
        rideId:    { type: string, format: uuid }
        score:     { type: integer, minimum: 1, maximum: 5 }
        comment:   { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
...